Started with JS
had to adjust measurements and other lines of code for it to look and function right. with the help of chatgpt, i was able to get it right. the x icon was really unnecessarily challenging. it's supposed to automatically look like an x as showed in the tutorial and chatgpt as well but i looked at it differently and changed the size of the lines instead and now it kinda looks proper.